Create SpeechRecognitionManager
This is SpeechRecognitionManager.swift which goes in the unity Plugins/iOS folder in Unity.  In xcode make sure to also add to build phases 'link binary with libraries' UnityFramework.framework and also Speech.framework, and in info.plist add the privacy speechrecognition and microphone.  Under 'General' add Speech.framework and UnityFramework.framework